Few little touch-ups for CD, ToG, Echoes ACT 2, The Lock (buffed)

+++ Projectiles/PlayerStands/Echoes/EchoesStandT3.cs
@@ -28,12 +28,16 @@ namespace JoJoStands.Projectiles.PlayerStands.Echoes
 
         private int ACT = 2;
         private int changeActCooldown = 20;
+        private int shoot = 0;
+        private int holdSpecial = 0;
+        private int echoesTailTipType = 1;
 
         private float remoteRange = 1200f;
 
         private bool remoteMode = false;
         private bool mouseControlled = false;
         private bool returnToPlayer = false;
+        private bool returnTail = false;
         private bool changeACT = false;
         private bool evolve = false;
 
@@ -50,6 +54,8 @@ namespace JoJoStands.Projectiles.PlayerStands.Echoes
             SelectAnimation();
             UpdateStandInfo();
             UpdateStandSync();
+            if (shoot > 0)
+                shoot--;
             if (shootCount > 0)
                 shootCount--;
             if (changeActCooldown > 0)
@@ -67,11 +73,6 @@ namespace JoJoStands.Projectiles.PlayerStands.Echoes
 
             Projectile.tileCollide = true;
 
-            Rectangle rectangle = Rectangle.Empty;
-
-            if (Projectile.owner == player.whoAmI)
-                rectangle = new Rectangle((int)(Main.MouseWorld.X), (int)(Main.MouseWorld.Y), 20, 20);
-
             if (mPlayer.usedEctoPearl && remoteRange == 1200f)
                 remoteRange *= 1.5f;
 
@@ -79,7 +80,7 @@ namespace JoJoStands.Projectiles.PlayerStands.Echoes
             {
                 if (!remoteMode)
                 {
-                    if (Main.mouseLeft && Projectile.owner == Main.myPlayer && !mPlayer.posing && !returnToPlayer)
+                    if (Main.mouseLeft && Projectile.owner == Main.myPlayer && !mPlayer.posing && !returnToPlayer && !returnTail && shoot == 0)
                     {
                         Punch();
                     }
@@ -88,15 +89,79 @@ namespace JoJoStands.Projectiles.PlayerStands.Echoes
                         if (player.whoAmI == Main.myPlayer)
                             attackFrames = false;
                     }
-                    if (Main.mouseRight && Projectile.owner == Main.myPlayer && !player.HasBuff(ModContent.BuffType<AbilityCooldown>()) && !returnToPlayer) //right-click ability 
+                    if (Main.mouseRight && Projectile.owner == Main.myPlayer && player.ownedProjectileCounts[ModContent.ProjectileType<EchoesTailTip>()] == 0 && !returnToPlayer && shoot == 0 && !returnTail && shootCount == 0) //right-click ability 
+                        holdSpecial++;
+                    if (!Main.mouseRight && Projectile.owner == Main.myPlayer && player.ownedProjectileCounts[ModContent.ProjectileType<EchoesTailTip>()] == 0 && !returnToPlayer && shoot == 0 && !returnTail && holdSpecial > 0 && holdSpecial < 60 && shootCount == 0)
                     {
-
+                        holdSpecial = 0;
+                        if (echoesTailTipType == 1 && shootCount == 0)
+                        {
+                            echoesTailTipType = 2;
+                            Main.NewText("KABOOM", Color.DeepSkyBlue);
+                            shootCount = 30;
+                        }
+                        if (echoesTailTipType == 2 && shootCount == 0)
+                        {
+                            echoesTailTipType = 3;
+                            Main.NewText("WOOOSH", Color.LightSkyBlue);
+                            shootCount = 30;
+                        }
+                        if (echoesTailTipType == 3 && shootCount == 0)
+                        {
+                            echoesTailTipType = 4;
+                            Main.NewText("SIZZLE", Color.IndianRed);
+                            shootCount = 30;
+                        }
+                        if (echoesTailTipType == 4 && shootCount == 0)
+                        {
+                            echoesTailTipType = 1;
+                            Main.NewText("BOING", Color.HotPink);
+                            shootCount = 30;
+                        }
                     }
-                    if (!attackFrames && !returnToPlayer)
+                    if (holdSpecial >= 60)
+                    {
+                        holdSpecial = 0;
+                        Projectile.frame = 0;
+                        shoot += 60;
+                        Vector2 shootVel = Main.MouseWorld - Projectile.Center;
+                        if (shootVel == Vector2.Zero)
+                            shootVel = new Vector2(0f, 1f);
+                        shootVel.Normalize();
+                        shootVel *= 8f;
+                        int proj = Projectile.NewProjectile(Projectile.GetSource_FromThis(), Projectile.Center, shootVel, ModContent.ProjectileType<EchoesTailTip>(), (int)(AltDamage * mPlayer.standDamageBoosts), 6f, Projectile.owner, Projectile.whoAmI);
+                        Main.projectile[proj].GetGlobalProjectile<JoJoGlobalProjectile>().echoesTailTipTier = mPlayer.echoesTier;
+                        Main.projectile[proj].GetGlobalProjectile<JoJoGlobalProjectile>().echoesTailTipType = echoesTailTipType;
+                        Main.projectile[proj].netUpdate = true;
+                        Projectile.netUpdate = true;
+                    }
+                    if (!attackFrames && !returnToPlayer && !returnTail)
                         StayBehind();
                 }
 
-                if (SpecialKeyPressedNoCooldown() && Projectile.owner == Main.myPlayer && !returnToPlayer) //remote mode
+                if (mPlayer.echoesTailTip != -1 && shoot == 0)
+                {
+                    if (Main.mouseRight && Projectile.owner == Main.myPlayer && !remoteMode && !returnToPlayer && !returnTail)
+                    {
+                        if (Main.projectile[mPlayer.echoesTailTip].GetGlobalProjectile<JoJoGlobalProjectile>().echoesTailTipStage != 2)
+                        {
+                            shootCount = 30;
+                            Main.projectile[mPlayer.echoesTailTip].GetGlobalProjectile<JoJoGlobalProjectile>().echoesTailTipStage = 2;
+                        }
+                        if (Main.projectile[mPlayer.echoesTailTip].GetGlobalProjectile<JoJoGlobalProjectile>().echoesTailTipStage == 2 && shootCount == 0 && Vector2.Distance(Projectile.Center, Main.projectile[mPlayer.echoesTailTip].Center) <= 1200f)
+                        {
+                            shootCount = 30;
+                            returnTail = true;
+                        }
+                        if (Main.projectile[mPlayer.echoesTailTip].GetGlobalProjectile<JoJoGlobalProjectile>().echoesTailTipStage == 2 && shootCount == 0 && Vector2.Distance(Projectile.Center, Main.projectile[mPlayer.echoesTailTip].Center) > 1200f)
+                        {
+                            shootCount = 30;
+                            Main.NewText("Out of reach");
+                        }
+                    }
+                }
+
+                if (SpecialKeyPressedNoCooldown() && Projectile.owner == Main.myPlayer && !returnToPlayer && !returnTail && shoot == 0) //remote mode
                 {
                     remoteMode = !remoteMode;
                     if (remoteMode)
@@ -105,28 +170,19 @@ namespace JoJoStands.Projectiles.PlayerStands.Echoes
                         Main.NewText("Remote Mode: Disabled");
                 }
 
-                if (SecondSpecialKeyPressedNoCooldown() && Projectile.owner == Main.myPlayer && changeActCooldown == 0 && mPlayer.echoesTier > 2)
+                if (SecondSpecialKeyPressedNoCooldown() && Projectile.owner == Main.myPlayer && changeActCooldown == 0 && mPlayer.echoesTier > 2 && !evolve)
                 {
                     changeACT = true;
                     Projectile.Kill();
                 }
 
-                if (Vector2.Distance(player.Center, Projectile.Center) <= remoteRange * 0.9f)
-                {
-                    if (remoteMode || returnToPlayer)
-                    {
-                        if (Projectile.velocity.X > 0)
-                            Projectile.spriteDirection = 1;
-                        if (Projectile.velocity.X < 0)
-                            Projectile.spriteDirection = -1;
-                    }
-                }
-                if (remoteMode && !returnToPlayer)
+                if (remoteMode && !returnToPlayer && !returnTail)
                 {
                     float distance = Vector2.Distance(Projectile.Center, Main.MouseWorld);
                     float halfScreenWidth = (float)Main.screenWidth / 2f;
                     float halfScreenHeight = (float)Main.screenHeight / 2f;
                     mPlayer.standRemoteModeCameraPosition = Projectile.Center - new Vector2(halfScreenWidth, halfScreenHeight);
+                    holdSpecial = 0;
                     Projectile.rotation = Projectile.velocity.X * 0.05f;
                     if (Main.mouseLeft && Projectile.owner == Main.myPlayer && !mPlayer.posing)
                     {
@@ -165,27 +221,71 @@ namespace JoJoStands.Projectiles.PlayerStands.Echoes
 
                     LimitDistance(remoteRange);
                 }
-
-                if (mPlayer.standAutoMode && !returnToPlayer) //automode
-                {
-                    remoteMode = false;
-                    BasicPunchAI();
-                }
-
-                if (Projectile.Distance(player.Center) >= newMaxDistance + 10f && !returnToPlayer && !remoteMode) //if suddenly stand is too far
-                    returnToPlayer = true;
-                if (Projectile.Distance(player.Center) <= 20f)
-                    returnToPlayer = false;
-
-                if (returnToPlayer)
-                {
-                    Projectile.tileCollide = false;
-                    MovementAI(player.Center, 8f + player.moveSpeed * 2);
-                }
-
-                if (player.teleporting)
-                    Projectile.position = player.position;
             }
+
+            if (Vector2.Distance(player.Center, Projectile.Center) <= remoteRange * 0.9f)
+            {
+                if (remoteMode || returnToPlayer || returnTail)
+                {
+                    if (Projectile.velocity.X > 0)
+                        Projectile.spriteDirection = 1;
+                    if (Projectile.velocity.X < 0)
+                        Projectile.spriteDirection = -1;
+                }
+            }
+
+            if (mPlayer.echoesTailTip != -1)
+            {
+                if (returnTail)
+                    MovementAI(Main.projectile[mPlayer.echoesTailTip].Center, 8f);
+                if (Main.projectile[mPlayer.echoesTailTip].GetGlobalProjectile<JoJoGlobalProjectile>().echoesTailTipStage == 2 && Vector2.Distance(Projectile.Center, Main.projectile[mPlayer.echoesTailTip].Center) <= 20f)
+                    Main.projectile[mPlayer.echoesTailTip].Kill();
+            }
+
+            if (mPlayer.echoesTailTip == -1 && returnTail)
+            {
+                returnTail = false;
+                returnToPlayer = true;
+            }
+
+            if (mPlayer.standAutoMode && !returnToPlayer && !returnTail) //automode
+            {
+                holdSpecial = 0;
+                remoteMode = false;
+                BasicPunchAI();
+            }
+
+            if (Projectile.Distance(player.Center) >= newMaxDistance + 10f && !returnToPlayer && !remoteMode && !returnTail) //if suddenly stand is too far
+                returnToPlayer = true;
+            if (Projectile.Distance(player.Center) <= 20f)
+                returnToPlayer = false;
+
+            if (returnToPlayer)
+            {
+                holdSpecial = 0;
+                Projectile.tileCollide = false;
+                MovementAI(player.Center, 8f + player.moveSpeed * 2);
+            }
+
+            if (returnTail)
+            {
+                Projectile.tileCollide = false;
+                holdSpecial = 0;
+            }
+
+            if (player.teleporting)
+            {
+                holdSpecial = 0;
+                Projectile.position = player.position;
+                returnTail = false;
+            }
+
+            if (player.HasBuff(ModContent.BuffType<StrongWill>()) && mPlayer.echoesTier == 3 && mPlayer.echoesACT3Evolve >= 20000 && Main.hardMode && mPlayer.echoesTailTip != -1)
+            {
+                evolve = true;
+                Projectile.Kill();
+            }
+
         }
 
         private void MovementAI(Vector2 target, float speed)
@@ -222,13 +322,20 @@ namespace JoJoStands.Projectiles.PlayerStands.Echoes
                 attackFrames = false;
                 PlayAnimation("Pose");
             }
+            if (shoot > 0)
+            {
+                idleFrames = false;
+                attackFrames = false;
+                PlayAnimation("Shoot");
+            }
+
         }
 
         public override void PlayAnimation(string animationName)
         {
             MyPlayer mPlayer = Main.player[Projectile.owner].GetModPlayer<MyPlayer>();
             string pathAddition = "";
-            if (Main.player[Projectile.owner].HasBuff(ModContent.BuffType<AbilityCooldown>()))
+            if (Main.player[Projectile.owner].ownedProjectileCounts[ModContent.ProjectileType<EchoesTailTip>()] > 0)
                 pathAddition = "Tailless_";
             if (Main.netMode != NetmodeID.Server)
                 standTexture = GetStandTexture("JoJoStands/Projectiles/PlayerStands/Echoes", "/EchoesACT2_" + pathAddition + animationName);
@@ -253,23 +360,27 @@ namespace JoJoStands.Projectiles.PlayerStands.Echoes
         public override void SendExtraStates(BinaryWriter writer)
         {
             writer.Write(ACT);
+            writer.Write(shoot);
 
             writer.Write(remoteRange);
 
             writer.Write(remoteMode);
             writer.Write(mouseControlled);
             writer.Write(returnToPlayer);
+            writer.Write(returnTail);
         }
 
         public override void ReceiveExtraStates(BinaryReader reader)
         {
             ACT = reader.ReadInt32();
+            shoot = reader.ReadInt32();
 
             remoteRange = reader.ReadSingle();
 
             remoteMode = reader.ReadBoolean();
             mouseControlled = reader.ReadBoolean();
             returnToPlayer = reader.ReadBoolean();
+            returnTail = reader.ReadBoolean();
         }
 
         public override void Kill(int timeLeft)
@@ -301,6 +412,7 @@ namespace JoJoStands.Projectiles.PlayerStands.Echoes
             if (evolve)
             {
                 player.maxMinions += 1;
+                mPlayer.echoesACT3Evolve = 0;
                 mPlayer.StandSlot.SlotItem.type = ModContent.ItemType<EchoesACT3>();
                 mPlayer.StandSlot.SlotItem.SetDefaults(ModContent.ItemType<EchoesACT3>());
                 Projectile.NewProjectile(Projectile.GetSource_FromThis(), Projectile.position, Projectile.velocity, ModContent.ProjectileType<EchoesStandFinal>(), 0, 0f, Main.myPlayer, 2f);
